Make sure chats and messages have been fetched...
before creating new ones.

With the latest Backbone.browserStorage, all operations are now
asynchronous.

When doing a read operation on a collection (due to a `fetch`) and
then immediately creating a model in that collection, then once the read
operation finishes (after creating the model), the collection is emptied
(e.g. when no results are returned from the fetch).

The solution is to first wait for the fetches to complete.

Did this now with messages and chatboxes, might be necessary in other
cases as well.

updates #1105